# Conflicts:
#	src/CMakeLists.txt
#	src/main.cpp

+++ src/player/schaden/schaden.h
@@ -8,8 +8,8 @@ public:
     int Gethealth() const { return hp; }
 
     // Health
-    int hp;
-    int maxHp;
+    float hp;
+    float maxHp;
     bool takeDamage;
     // Damage cooldown
     float invincibleTimer;
